minor bug fixes to grid_opt.py
args.optparm is created , but then used as args.optparam, so I changed it to args.optparm
args.na_fill is never defined. I added a line to provide a default value for na_fill. Probably needs to be a command line argument with a default value, but this at least allows the code to run.