Rewrite HSI calcultation
This commit adds a script to calculate the habitat suitability index. It
is basically a rewritten version of `hsi_Bever_updated.py` using xarray
(or rather uxarray/suxarray). The script needs the same inputs used for
the original script. Outputs are all in NetCDF.
Note that this uses the main HEAD of suxarray and uxarray ec7c8c057b1
commit.

Closes SCH-507