Add a script to combine hotstart files
This commit adds a bash script to combine hotstart files in outputs directory with a user input and create a symbolic link. Note that the script needs to run in a study directory. The script assumes many items such as hotstart output intervals and dt, so they need to be modified according to a simulation setting.

SCH-726