Provide better control for preventing overwriting
Previously, uv3D.th.nc was renamed with a timestamp if it already existed in fg_dir.
Now, the user can choose to overwrite it or raise an error.
Also, hard-wired file name "uv3D.th.nc" is replaced with a user-defined name.

New function arguments are 'output_name' and 'overwrite_existing'.